fix(market-profile): address PR #618 round-2 review (layer move, test coverage, exact matchers)
B1: move vi.mock('@/shared/api/fmp/httpClient') above all imports in fmpCryptoListClient.test.ts (import/first ESLint fix)
B2: move fetchCryptoAssetList out of entities/ticker/lib/ (pure-only) into api.ts; update seed script import path; move fetchCryptoAssetList test into __tests__/api.test.ts with vi.mock at top
B3: replace inline import() type with top-level `import type { MarketProfileId }` from marketProfile/types (not barrel) to avoid circular dependency
B4: add 'falls back to symbol when name is absent' test for mapCryptoListRow
B5: replace .toContain()/.not.toContain() with exact .toEqual() on us-equity.tabs, crypto.tabs, crypto.allowedTimeframes
S1: expand US_EXCHANGES JSDoc in usEquity.ts to note fmpTickerApi mirror + sync requirement (MISTAKES.md §16.5)
S2: extract PriceFormatConfig interface from inline priceFormat type in types.ts; export from barrel index.ts
